Fixed issue with file_mover function
I had incorrectly set up a test in the file_mover function that was supposed to determine whether or not the user entered a source and destination folder.  I corrected this by initializing variables in the class for holding data pertaining to the source and destination folder and implementing a check on the values stored in those variables.  I also noticed that file_mover threw an error if a file that was to be moved was already in the destination folder.  I introduced a short subroutine to compare files in the source folder with files in the destination folder and removing files that were already present in both folders before calling shutil().  This seemed to resolve that problem.